Create ConnorLariviere-A3Q1.js
In this assignment question, we created a catcher sprite for the 'hungry bird' game in Assignment 3. I chose to make the sprite a UFO that is put at the top of the screen. It moves with your mouse position. The link to the original sketch is here: https://editor.p5js.org/connorL/sketches/d-MvOPE50