Create ConnorLariviere-A3Q2.js
In this assignment question, we made the sprite react to the mouse being pressed. In my case, the UFO's lights turned on and it's super hi-tech tractor beam activated.
Also, I added background elements, such as the moon and a grassy field. The link to the original sketch is here: https://editor.p5js.org/connorL/sketches/slPEesTef