Create ConnorLariviere-A3Q5.js
This is the fully finished game for Assignment 5. New additions are that the cow rises faster if the tractor beam is on and the cow is below the UFO, the cows now disappear when colliding with the UFO and they increment the score, which is shown. 
The link to the sketch is here: https://editor.p5js.org/connorL/sketches/l7sh0DaJf